# apps/command/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {
      "name": "Team VIT, SIH 2026 PS SIH26085"
    },
    "description": "Street-level urban flood nowcasting digital twin (SIH 2026, PS SIH26085, MoES). Every response carries `run_id` and `valid_ts` where a run is involved; errors use `{ error: { code, message, run_id } }`. Endpoints answering 501 are contract-only until their engine's phase lands.",
    "title": "VARUNA API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/route": {
      "post": {
        "description": "Plan one trip.\n\nBody: ``{origin, destination, depart_at?, profile?, risk_tolerance?, run_id?, spread?,\ntrip_id?, explain?}``.\n\nReturns both routes, because the comparison is the product: a dispatcher who only sees the\nsafe route has no way to judge whether the detour was worth it.\n\n``spread`` (default true) adds ``corridors[]`` - up to three safe roads with the share of\ntraffic the policy gives each, and the one this request is assigned to. ``trip_id`` makes\nthat assignment stable for a trip, so a reader who reloads is not sent somewhere else; it is\nthe client's own random id and nothing is stored against it. ``explain`` (default true) adds\n``reasons[]``, which is structured data - the frontend writes the sentences (TECH_SPEC 3.2).",
        "operationId": "route_v1_route_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Route V1 Route Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Route around the forecast water, beside what a naive router does",
        "tags": [
          "route"
        ]
      }
    },
...
    "/v1/verification": {
      "get": {
        "description": "Detection, timing and the threshold sweep, plus what could not be scored and why.",
        "operationId": "verification_v1_verification_get",
        "parameters": [
          {
            "description": "Bundle id, e.g. MUM-2019-07-02",
            "in": "query",
            "name": "event",
            "required": false,
            "schema": {
              "default": "MUM-2019-07-02",
              "description": "Bundle id, e.g. MUM-2019-07-02",
              "title": "Event",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Verification V1 Verification Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scores for an event against its sourced ground truth",
        "tags": [
          "verification"
        ]
      }
    },
    "/v1/verification/rain-skill": {
      "get": {
        "description": "Rain CSI, POD and FAR at 10, 20 and 40 mm/h by lead, pooled over the event's baked cycles.\n\nScored for the ensemble mean, the ensemble median and persistence (the analysis each nowcast\nstarted from, held), with the Brier score and reliability of the members' exceedance\nprobability, the sample size behind every lead and the useful skill horizon it computes. An\nevent whose bundle has no truth field or no baked rain answers ``available: false`` with the\nreason; an unknown event is a 404.",
        "operationId": "verification_rain_skill_v1_verification_rain_skill_get",
        "parameters": [
          {
            "description": "Bundle id, e.g. MUM-2019-07-02",
            "in": "query",
            "name": "event",
            "required": false,
            "schema": {
              "default": "MUM-2019-07-02",
              "description": "Bundle id, e.g. MUM-2019-07-02",
              "title": "Event",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Verification Rain Skill V1 Verification Rain Skill Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rain skill by lead time against the bundle's reconstructed truth field",
        "tags": [
          "verification"
        ]
      }
    },
...
  },
...
}